Finalizing for presentation

+++ src/main/java/hu/unideb/inf/todo/service/TodoService.java
@@ -2,14 +2,22 @@ package hu.unideb.inf.todo.service;
 
 import hu.unideb.inf.todo.dto.todo.TodoDTO;
 import hu.unideb.inf.todo.exception.todo.TodoNotFoundException;
+import hu.unideb.inf.todo.exception.user.JWTInvalidException;
+import hu.unideb.inf.todo.exception.user.NoSuchUserException;
 import hu.unideb.inf.todo.model.Todo;
+import hu.unideb.inf.todo.model.UserModel;
 import hu.unideb.inf.todo.repository.TodoRepository;
+import hu.unideb.inf.todo.security.JwtTokenProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
+
+import javax.servlet.http.HttpServletRequest;
 
 @Service
 public class TodoService {
@@ -17,13 +25,27 @@ public class TodoService {
     @Autowired
     private TodoRepository todoRepository;
 
-    public Iterable<Todo> getAllTodo() {
-        return todoRepository.findAll();
+    @Autowired
+    JwtTokenProvider jwtTokenProvider;
+
+    @Autowired
+    UserService userService;
+
+    public Iterable<Todo> getAllTodo(HttpServletRequest req) {
+        Optional<UserModel> user = userService.getUserByUsername(jwtTokenProvider.getUsername(jwtTokenProvider.resolveToken(req)));
+        if (user.isEmpty()) {
+            throw new JWTInvalidException();
+        }
+        return todoRepository.findTodosByUserId(user.get().getId());
     }
 
     @Transactional
-    public Todo getTodoById(long id) {
-        Optional<Todo> todo = todoRepository.getTodoById(id);
+    public Todo getTodoById(HttpServletRequest req, long id) {
+        Optional<UserModel> user = userService.getUserByUsername(jwtTokenProvider.getUsername(jwtTokenProvider.resolveToken(req)));
+        if (user.isEmpty()) {
+            throw new JWTInvalidException();
+        }
+        Optional<Todo> todo = todoRepository.findTodoByIdAndUserId(id, user.get().getId());
         if (todo.isPresent()) {
             return todo.get();
         } else {
@@ -32,32 +54,47 @@ public class TodoService {
     }
 
     @Transactional
-    public Todo newTodo(TodoDTO todoDTO) {
+    public Todo newTodo(HttpServletRequest req, TodoDTO todoDTO) {
         Todo todo = new Todo();
+
+        Optional<UserModel> user = userService.getUserByUsername(jwtTokenProvider.getUsername(jwtTokenProvider.resolveToken(req)));
+        if (user.isEmpty()) {
+            throw new JWTInvalidException();
+        }
         todo.setContent(todoDTO.getContent());
+        todo.setUser(user.get());
         return todoRepository.save(todo);
     }
 
     @Transactional
     @Modifying
-    public Todo deleteTodoById(long id) {
-        Todo todo = todoRepository.getTodoById(id).orElseThrow(() -> new TodoNotFoundException(id));
+    public Todo deleteTodoById(HttpServletRequest req, long id) {
+        Optional<UserModel> user = userService.getUserByUsername(jwtTokenProvider.getUsername(jwtTokenProvider.resolveToken(req)));
+        if (user.isEmpty()) {
+            throw new JWTInvalidException();
+        }
+        Optional<Todo> optionalTodo = todoRepository.findTodoByIdAndUserId(id, user.get().getId());
+        if (optionalTodo.isEmpty()) {
+            throw new TodoNotFoundException(id);
+        }
         todoRepository.deleteTodoById(id);
-        return todo;
+        return optionalTodo.get();
     }
 
     @Transactional
-    public Todo updateTodo(TodoDTO todoDTO, long id) {
-        try {
-            Todo todo = getTodoById(id);
-            todoRepository.updateTodo(todoDTO.getContent(), id);
-            todo.setContent(todoDTO.getContent());
-            return todo;
-        } catch (TodoNotFoundException ex) {
-            Todo newTodo = new Todo();
-            newTodo.setContent(todoDTO.getContent());
-            return todoRepository.save(newTodo);
+    public Todo updateTodo(HttpServletRequest req, TodoDTO todoDTO, long id) {
+        Optional<UserModel> user = userService.getUserByUsername(jwtTokenProvider.getUsername(jwtTokenProvider.resolveToken(req)));
+        if (user.isEmpty()) {
+            throw new JWTInvalidException();
         }
+        Optional<Todo> optionalTodo = todoRepository.findTodoByIdAndUserId(id, user.get().getId());
+        if (optionalTodo.isEmpty()) {
+            throw new TodoNotFoundException(id);
+        }
+        Todo todo = optionalTodo.get();
+        todoRepository.updateTodo(todoDTO.getContent(), id);
+        todo.setContent(todoDTO.getContent());
+        return todo;
     }
 
 }
